Add files via upload
The purpose of this project
is to develop a sophisticated system for the detection of credit card fraudulent activities. With the increasing use of technology, credit card fraud has become a major concern for consumers and financial institutions, resulting in significant financial losses. To address this issue, it is essential to have systems in place that can detect fraudulent activities and minimize losses.

One of the key challenges in detecting credit card fraud is the imbalance in the distribution of fraudulent and non-fraudulent transactions. Fraudulent transactions constitute a small fraction of all transactions, making it difficult to detect them using traditional methods.

To overcome this challenge, I have obtained a credit card fraudulent activities dataset from Kaggle. The dataset will be used to analyze the relationships between various features and detect fraudulent transactions. Through exploratory data analysis, I aim to gain a deeper understanding of the dataset and develop a robust model for the detection of fraudulent activities. The link to the dataset has been providedt: https://www.kaggle.com/mlg-ulb/creditcardfraud

Goal:
The goal of this project is to use machine learning techniques to improve the accuracy of detecting credit card fraudulent activities. The model will be trained on a dataset of credit card transactions and the goal is to improve the measure called Area Under the Precision-Recall Curve (AUPRC). Additionally, the project aim to acquire more knowledge about the financial industry and how to apply machine learning to real-world problems